requirements.txt
absl-py==0.9.0
alabaster==0.7.12
altair==4.1.0
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.7
astor==0.8.1
astroid==2.4.2
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
base58==2.0.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
blinker==1.4
boto3==1.14.2
botocore==1.17.2
bottle==0.12.18
bottle-websocket==0.2.9
Brotli==1.0.7
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
confuse==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
cssutils==1.0.2
cycler==0.10.0
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
decorator==4.4.2
decouple==0.0.7
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
Eel==0.12.4
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Compress==1.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
gast==0.3.3
get==2019.4.13
gevent==20.6.1
gevent-websocket==0.10.1
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.16
grpcio==1.29.0
h5py==2.10.0
helpdev==0.7.1
html5lib==1.0.1
htmlmin==0.1.12
httpie==2.1.0
idna==2.9
ImageHash==4.1.0
imagesize==1.2.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.15.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyboard==0.13.5
keyring==21.2.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.33.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mpld3==0.3
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.50.0
numpy==1.18.5
numpydoc==1.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pandas-profiling==2.8.0
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pefile==2019.4.18
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.8.1
pluggy==0.13.1
post==2019.4.13
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
psycopg2==2.8.5
ptpython==3.0.2
ptyprocess==0.6.0
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.4.0b2
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
PyInstaller==3.6
pylint==2.5.3
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.4
QtPy==1.9.0
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.1.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
streamlit==0.61.0
sympy==1.6
tabulate==0.8.7
tangled-up-in-unicode==0.0.6
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
Theano==1.0.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
tqdm==4.46.1
traitlets==4.3.3
ttp==0.3.0
tweepy==3.8.0
typed-ast==1.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.24.3
validators==0.15.0
vega==3.4.0
virtualenv==20.0.21
visions==0.4.4
watchdog==0.10.2
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
yapf==0.30.0
youtube-dl==2020.6.6
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
